refactor: Moved similar missing components together
