+++ src/pages/models/index.js
@@ -2,13 +2,13 @@ import React, { useEffect } from 'react'
 import useApp from '../../hooks/useApp'
 import withLanguage from '../../components/withLanguage'
 import AppWrapper from '../../components/app/wrapper'
-import CenteredLayout from '../../components/layouts/centered'
+import WideLayout from '../../components/layouts/wide'
 
 import { Link } from 'gatsby'
 import { FormattedMessage } from 'react-intl'
 import Button from '@material-ui/core/Button'
 import Avatar from '../../components/avatar'
-import NoModel from '../../components/no-model'
+import MissingModels from '../../components/missing/models'
 import ModelGraph from '../../components/model-graph.js'
 
 const ModelsIndexPage = props => {
@@ -43,41 +43,43 @@ const ModelsIndexPage = props => {
 
   return (
     <AppWrapper app={app}>
-      <CenteredLayout app={app}>
+      <WideLayout app={app} top>
         {Object.keys(app.models).length > 0 ? (
-          Object.keys(app.models).map((handle, model) => {
-            return (
-              <div className="box" key={handle}>
-                <Link
-                  to={'/models/' + handle}
-                  title={app.models[handle].name}
-                  data-test="model-link"
-                >
-                  <div style={styles.model}>
-                    <div style={styles.avatar}>
-                      <Avatar variant="model" data={app.models[handle]} />
+          <>
+            {Object.keys(app.models).map((handle, model) => {
+              return (
+                <div className="box" key={handle}>
+                  <Link
+                    to={'/models/' + handle}
+                    title={app.models[handle].name}
+                    data-test="model-link"
+                  >
+                    <div style={styles.model}>
+                      <div style={styles.avatar}>
+                        <Avatar variant="model" data={app.models[handle]} />
+                      </div>
+                      <div>
+                        <h6 style={styles.name}>{app.models[handle].name}</h6>
+                        <p style={styles.notes}>{app.models[handle].notes}</p>
+                      </div>
                     </div>
                     <div>
-                      <h6 style={styles.name}>{app.models[handle].name}</h6>
-                      <p style={styles.notes}>{app.models[handle].notes}</p>
+                      <ModelGraph model={app.models[handle]} intl={app.intl} />
                     </div>
-                  </div>
-                  <div>
-                    <ModelGraph model={app.models[handle]} intl={app.intl} />
-                  </div>
-                </Link>
-              </div>
-            )
-          })
+                  </Link>
+                </div>
+              )
+            })}
+            <p style={{ textAlign: 'right' }}>
+              <Button size="large" variant="contained" color="primary" href="/model">
+                <FormattedMessage id="app.newModel" />
+              </Button>
+            </p>
+          </>
         ) : (
-          <NoModel />
+          <MissingModels />
         )}
-        <p style={{ textAlign: 'right' }}>
-          <Button size="large" variant="contained" color="primary" href="/model">
-            <FormattedMessage id="app.newModel" />
-          </Button>
-        </p>
-      </CenteredLayout>
+      </WideLayout>
     </AppWrapper>
   )
 }
